Add files via upload
Task 1: Basic Agent with Tool
Objective
The objective of this task is to construct a basic agent and equip it with a custom tool. The agent should effectively utilize the tool to perform a specified function. You can use the agentic framework of your choice (LangGraph, AutoGen, CrewAI etc), or go ahead without choosing a specialized framework. We recommend using AutoGen/LangGraph for those without any specific preference.
We understand if you haven’t used any of the above mentioned frameworks before. This task is meant for you to explore and get accustomed to Agentic Frameworks, and demonstrate your ability to pick up new frameworks quickly.
Methodology
1. Agent Development
Build a rudimentary multi-agent application capable of executing predefined tasks.
2. Tool Creation
Develop a simple tool using the yfinance API to fetch financial data such as stock prices, and visualize it.
3. Integration and Demonstration
Integrate the tool with the agent.
Provide specific instructions to the agent and showcase its ability to use the tool effectively.
Deliverables
Think of a particular use case, and implement a basic agent for it. Clearly state what your agent is capable of doing.
A functional tool built using the yfinance API. Feel free to use any other API that better suits your use case.
A demonstration of the agent's ability to utilize the tool effectively. Feel free to use as many tools as you like. Refrain from using existing implementations available on the internet, We want to see something cool!